Spool bucket cleanups:
   There can only be one spool bucket for a given tempfile,
because that bucket marks where new data may be appended.
So fo now we mark copy() as unimplemented (may fix this later
once we determine whether the orignal, or the copy, keeps 
the spool tag). Now split() is corrected to convert
the lead bucket back to an ordinary apr_bucket_type_file.


git-svn-id: https://svn.apache.org/repos/asf/httpd/apreq/branches/multi-env-unstable@151415 13f79535-47bb-0310-9956-ffa450edef68